requirements.txt
numpy
pyyaml
scikit-learn
pytest
open3d
torch
torchvision
torchaudio